Reorganize tests for bookmarks into one file
